requirements.txt
pystache==0.6.5
PyYAML==6.0.1
watchdog==4.0.0
